chromium/third_party/blink/web_tests/virtual/legacy-client-hints-no-fp-delegation/external/wpt/client-hints/accept-ch-stickiness/http-equiv-same-origin-iframe.https-expected.txt

This is a testharness.js-based test.
PASS http-equiv same origin iframe precondition: Test that the browser does not have client hints preferences cached
PASS http-equiv same origin iframe set Accept-CH
FAIL http-equiv same origin iframe got client hints according to expectations. assert_equals: message from opened page expected "PASS" but got "FAIL"
Harness: the test ran to completion.
